power_supply: Enable Battery temp zone change interrupt
Enable Battery Temperature  zone change interrupt feature
for Whiskey Cove PMIC.

To enable the Battery Zone change interrupt PMIC battery
Zone registers need to be configured with ADC values for
each zone. The ADC values are available in the LPAT table.
In existing implementation, the ADC lookup table is defined
in intel_soc_pmic_wc.c as a pdata for intel_pmic_ccsm driver.
This definition is replaced with a LPAT table which is a
kelvin to ADC raw value mapping table. The lookup function
also changed accordingly. The PMIC zone change interrupts
were not enabled by default. So unmaksed the interrupt to
trigger interrupt on zone change.

Change-Id: Ie2b640fc6c321c792c5247312e1730a925057d38
Signed-off-by: Jenny TC <jenny.tc@intel.com>