PMIC-BCU: Avoid generating VWARNB interrupts continuously
When the battery is low and High power UC are running, we get repetitive
interrupts for voltage spikes on Vsys due to repetitive crossing of
VWARNB threshold. Based on experiments conducted 30 sec duration is good
enough to get in normal state. Without this VWARNB interrupt will occur
continuously (for low and recovery).

Hence, masking the VWARNB interrupt for 30seconds to reduce repetitive
interrupts.

Change-Id: I8309553d878b32c4be2bd2b2fe63c8a6d8d0ba0e
Tracked-on: https://jira01.devtools.intel.com/browse/IMINAN-14181
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>